+++ perf/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 # `perf-utils` CHANGELOG
 
+## 1.0.0
+
+Initial (well done) release. Renames functions (increase() -> speedup() and times() -> timesFaster())!
+
 ## 0.1.0
 
 Initial release. A bit rushed, lacks examples, test suite, proper README, and more stuff.